+++ bark/python_wrapper/world/evaluation.cpp
@@ -78,12 +78,11 @@ void python_evaluation(py::module m) {
       .def("__repr__", [](const EvaluatorStepCount& g) {
         return "bark.core.world.evaluation.EvaluatorStepCount";
       });
-  py::class_<EvaluatorCaptureAgentStates, BaseEvaluator,
-             std::shared_ptr<EvaluatorCaptureAgentStates>>(m, "EvaluatorCaptureAgentStates")
-      .def(py::init<>())
-      .def("__repr__", [](const EvaluatorCaptureAgentStates& g) {
-        return "bark.core.world.evaluation.EvaluatorCaptureAgentStates";
-      });
-  
+
+  m.def("CaptureAgentStates", py::overload_cast<const World&>(
+    &CaptureAgentStates<World>));
+  m.def("CaptureAgentStates", py::overload_cast<const ObservedWorld&>(
+    &CaptureAgentStates<ObservedWorld>));
+
   python_ltl(m.def_submodule("ltl", "LTL Rules"));
 }
